Add composer.lock to .gitignore
composer.lock is ignored because libraries do not need to add it to
version control. It might be needed to add it to version control when
the code should be published to the WordPress SVN (because the vendor
directory has to be included as well) to ensure that the exact same
dependencies are published.